+++ src/client/js/index.js
@@ -1,11 +1,12 @@
 /*
- * pwix:roles/src/client/js/index.js
+ * pwix:roles/src/client/js/accounts-ui.js
  */
 
 import '../../common/js/index.js';
 //
 import './ready.js';
 //
+import './accounts-ui.js';
 import './client.js';
 import './current.js';
 import './scopes.js';